Don't require relative paths to start with ./ or ../
Errors reported by Go, even when containing relative paths, do not
always begin with ./ or ../ – for example:

	$ go build ./...
	# honnef.co/go/tools/lintcmd
	lintcmd/format.go:28:8: syntax error: cannot use path := filepath.Clean(pos.Filename) as value

Related, linters commonly used with Go (such as Staticcheck) do not
prefix paths with ./ either.

This change slightly relaxes the stricter pattern that was introduced
as part of v2 to address #46. The pattern is still stricter than it
was in v1, and probably as strict as it can be.